requirements.txt
streamlit
langchain
langchain
qdrant-client
openai
pypdf
pypdf2
langchain-openai
langchain-community
uuid
tiktoken
pikepdf
python-dotenv
